Fix type conversion warning on win_pthread.h

+++ Headers/win_pthread.h
@@ -294,7 +294,7 @@ static inline int pthread_cond_timedwait (pthread_cond_t *cond, pthread_mutex_t 
   cond->nbwait++;
   winPthreadAssertPthread(pthread_mutex_unlock(mutex));
 again:
-  switch (WaitForSingleObject(cond->sem, time->tv_sec*1000+time->tv_nsec/1000)) {
+  switch (WaitForSingleObject(cond->sem, (DWORD) time->tv_sec*1000+time->tv_nsec/1000)) {
     default:
     case WAIT_FAILED:
       setSystemErrno();
